pyproject.toml
[project]
name = "xyz-a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "asyncio-atexit>=1.0.1",
    "dotenv>=0.9.9",
    "fastmcp>=2.3.4",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "pysocks>=1.7.1",
    "python-a2a>=0.5.5",
    "socksio>=1.0.0",
]
